requirements.txt
fastapi
uvicorn
black
pylint
fire
sqlalchemy
alembic
python-dotenv
psycopg2
passlib
bcrypt==4.0.1
python-multipart
python-jose
requests